requirements.txt
streamlit
torch
transformers
tensorflow
scikit-learn
joblib
pillow
spacy
numpy
accelerate
